requirements.txt
openai>=1.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
slack-sdk>=3.27.0
python-multipart>=0.0.9
sse-starlette>=2.1.0
pydantic>=2.7.0
